Checkpoint commit. I've got the asgi interface answering requests but now I've started to refactor the base interface code and I realized I should have a snapshot commit before I go crazy and rip a whole bunch of stuff out and consolidate some stuff. I think Router and Call can actually be part of the interface's BaseApplication and a lot of the methods can probably rely on datatypes functionality now
