--- endpoints/interface/uwsgi/__init__.py
@@ -1,11 +1,0 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals, division, print_function, absolute_import
-
-try:
-    import uwsgi
-except ImportError:
-    uwsgi = None
-
-from ...compat import *
-from ..wsgi import Application
-
